Switch to postgres

+++ db/migrate/20180129081511_add_score_to_bgame.rb
@@ -1,5 +1,5 @@
 class AddScoreToBgame < ActiveRecord::Migration[7.0]
   def change
-    add_column :bgames, :score, :double
+    add_column :bgames, :score, :float
   end
 end
